// src/model/ui/routes/AdminRoutesEnum.ts
export enum AdminRoutesEnum {
    SEMINARS = "/seminars",
    CALENDAR_MANAGEMENT = "/calendar-management",
    SEMINAR_CATS = "/seminar-cats",
    SEMINAR_SUB_CATS = "/sub-cats",
    EXAM_CATS = "/exam-cats",
    EXAM_SUB_CATS = "/sub-cats",
    BOOK_CATS = "/book-cats",
    BOOK_SUB_CATS = "/sub-cats",
    ACTIVITY_LOG = "/activity-log",
    INSTITUTIONS = "/institutions",
    FUNCTIONS = "/functions",
    ORGANIZERS = "/organizers",
    LOCATIONS = "/locations",
    DATA_VAULT = "/data-vault",
    EBL = "/electronic-book-library",
    USER_PROFILES = "/user-profiles",
    USER_CLASSES = "/user-classes",
    COST_CATS = "/cost-cats",
    COST_SUB_CATS = "/sub-cats",
    MODULES = "/modules",
    REPORT_DEFINITIONS = "/report-definitions",
    WORD_TEMPLATES = "/word-templates",
    EXAM_FORM = "/exam-form",
}