// src/model/ui/routes/MainRoutesEnum.ts
export enum MainRoutesEnum {
    HOME = "/",
    INTERNAL_PORTAL = "/internal-portal",
    INFORMATIONAL_PORTAL = "/informational-portal",
    E_LEARNING_PORTAL = "/e-learning-portal",
    SURVEY = "/s",
    JOURNAL_OF_CANDIDATES = "/internal-portal/the-journal-of-initial-candidates",
    ADMIN_PANEL = "/admin-panel",
    CALENDAR = "/calendar",
    ERROR_TEMPLATE = "/404",
    SUCCESS_TEMPLATE = "/200",
}