// src/model/domain/enums/AdminPanelPermissionEnum.ts
export enum AdminPanelPermissionEnum {
    ADMIN_DATA_VALUT = "ADMIN_DATA_VALUT",
    ADMIN_E_LIBRARY = "ADMIN_E_LIBRARY",
    ADMIN_ACTIVITY_LOG = "ADMIN_ACTIVITY_LOG",
    ADMIN_USER_PROFILES = "ADMIN_USER_PROFILES",
    ADMIN_USER_CLASSES = "ADMIN_USER_CLASSES",
    ADMIN_SEMINAR_CATS = "ADMIN_SEMINAR_CATS",
    ADMIN_BOOK_CATS = "ADMIN_BOOK_CATS",
    ADMIN_COST_CATS = "ADMIN_COST_CATS",
    ADMIN_EXAM_CATS = "ADMIN_EXAM_CATS",
    ADMIN_INSTITUTIONS = "ADMIN_INSTITUTIONS",
    ADMIN_FUNCTIONS = "ADMIN_FUNCTIONS",
    ADMIN_ORGANIZERS = "ADMIN_ORGANIZERS",
    ADMIN_LOCATIONS = "ADMIN_LOCATIONS",
    ADMIN_CLASS_MODULES = "ADMIN_CLASS_MODULES",
    ADMIN_REPORT_DEFINITONS = "ADMIN_REPORT_DEFINITONS",
    ADMIN_WORD_TEMPLATES = "ADMIN_WORD_TEMPLATES"
}